requirements.txt
torch==1.2.0
torchvision==0.4.0 -f https://download.pytorch.org/whl/torch_stable.html
numpy>=1.16.5
scipy>=1.3.0
opencv-python
gdown
pillow==6.2.2
gradio
